pulls out match table component from admin guest component
