+++ app/src/pages/AdminGuestView.tsx
@@ -9,6 +9,7 @@ import { faTimesCircle, faPaw, faSmokingBan, faWineBottle, faPrescriptionBottleA
 
 import { HostHomeType } from "../models/HostHomeType";
 import { AdminGuestStyle } from "./style"
+import MatchTable from '../components/Admin/Table'
 // import './AdminGuestView.css';
 
 const useStyles = makeStyles(theme => (
@@ -16,119 +17,12 @@ const useStyles = makeStyles(theme => (
         root: {
             flexGrow: 1
         },
-        toolbar: {
-            borderBottom: `1px solid ${theme.palette.divider}`,
-        },
-        toolbarTitle: {
-            flex: 1,
-        },
-        toolbarSecondary: {
-            justifyContent: 'space-between',
-            overflowX: 'auto',
-        },
-        toolbarLink: {
-            padding: theme.spacing(1),
-            flexShrink: 0,
-        },
-        mainFeaturedPost: {
-            position: 'relative',
-            backgroundColor: theme.palette.grey[800],
-            color: theme.palette.common.white,
-            marginBottom: theme.spacing(4),
-            backgroundSize: 'cover',
-            backgroundRepeat: 'no-repeat',
-            backgroundPosition: 'center',
-        },
-        overlay: {
-            position: 'absolute',
-            top: 0,
-            bottom: 0,
-            right: 0,
-            left: 0,
-            backgroundColor: 'rgba(0,0,0,.3)',
-        },
-        mainFeaturedPostContent: {
-            position: 'relative',
-            padding: theme.spacing(3),
-            [theme.breakpoints.up('md')]: {
-                padding: theme.spacing(6),
-                paddingRight: 0,
-            },
-        },
-        mainGrid: {
-            marginTop: theme.spacing(3),
-        },
-        card: {
-            display: 'flex',
-        },
-        cardDetails: {
-            flex: 1,
-        },
-        cardMedia: {
-            width: 160,
-        },
-        markdown: {
-            ...theme.typography.body2,
-            padding: theme.spacing(3, 0),
-        },
-        sidebarAboutBox: {
-            padding: theme.spacing(2),
-            backgroundColor: theme.palette.grey[200],
-        },
-        sidebarSection: {
-            marginTop: theme.spacing(3),
-        },
-        footer: {
-            backgroundColor: theme.palette.background.paper,
-            marginTop: theme.spacing(8),
-            padding: theme.spacing(6, 0),
-        },
         paper: {
             padding: theme.spacing(2),
             textAlign: 'center',
             color: theme.palette.text.secondary,
         },
-        table: {
-            minWidth: 650,
-        },
-        failedQuestion: {
-            fontWeight: 'bolder',
-            height: '100%',
-            width: '100%'
-        },
-        passedQuestion: {
-            backgroundColor: '#ecf0f1'
-        },
-        tableHeader: {
-            backgroundColor: '#00AAEF',
-            fontStyle: 'normal',
-            fontWeight: 'bold',
-            fontSize: '16px',
-            lineHeight: '22px',
-            color: '#FFFFFF'
-        },
-        tableHeaderCell: {
-            fontStyle: 'normal',
-            fontWeight: 'bold',
-            fontSize: '16px',
-            lineHeight: '22px',
-            color: '#FFFFFF',
-            textAlign: 'center'
-        },
-        matchingHeaderCell: {
-            fontStyle: 'normal',
-            fontWeight: 'bold',
-            fontSize: '16px',
-            lineHeight: '22px',
-            color: '#FFFFFF',
-            textAlign: 'center'
-        },
-        tableRowOdd: {
-            backgroundColor: '#FFFFFF'
-        },
-        tableRowEven: {
-            backgroundColor: '#F5F5F5'
-        }
+
     })));
 
 interface InterestDescription {
@@ -136,9 +30,6 @@ interface InterestDescription {
     lastUpdated: Date;
 }
 
-interface InterestMapping {
-    [key: number]: InterestDescription;
-}
 
 export const AdminGuestView = () => {
 
@@ -151,12 +42,6 @@ export const AdminGuestView = () => {
         guestQuestionsByKey
     } = useHostHomeData();
 
-    const petsKeys = ['pets_have', 'host_pets'];
-
-    const havePetsQuestion = guestQuestionsByKey.get('pets_have') as GuestQuestion;
-    const hostPetsQuestion = guestQuestionsByKey.get('host_pets') as GuestQuestion;
-
-    const history = useHistory();
 
     const guest: Guest = data.guests.find((g: Guest) => g.id === guestId) || {} as Guest;
 
@@ -193,38 +78,6 @@ export const AdminGuestView = () => {
         });
     }, [data.hosts, data.matchResults]);
 
-    // const unmatched = React.useMemo(() => {
-    //     return data.hosts.filter((host: Host) => matched.filter((matchedHost: Host) => host.id === matchedHost.id).length < 1)
-    // }, [data.matchResults]);
-
-    // const unmatched = React.useMemo(() => {
-    //     return data.hosts.filter((host: Host) => matched.filter((matchedHost: Host) => host.id === matchedHost.id).length < 1)
-    // }, [data.matchResults]);
-
-    const hostQuestionsFailed = data.matchResults
-        .filter((matchResult: MatchResult) => (
-            matchResult.guestId === guestId
-            && (matchResult.restrictionsFailed.length > 0 || matchResult.guestInterestLevel === GuestInterestLevel.NotInterested)
-        ))
-        .reduce<Map<number, Array<number>>>((prev: Map<number, Array<number>>, cur: MatchResult) => {
-            prev.set(cur.hostId, cur.restrictionsFailed.map((r: Restriction) => r.hostQuestionId));
-            return prev;
-        }, new Map<number, Array<number>>());
-
-    console.log(`hostQuestionsFailed: ${JSON.stringify(hostQuestionsFailed)}`);
-
-    const interestByHostId: InterestMapping = React.useMemo(() => {
-        return data.matchResults
-            .filter((matchResult: MatchResult) => matchResult.guestId === guestId)
-            .reduce<InterestMapping>((map: InterestMapping, matchResult: MatchResult, index: number) => {
-                map[matchResult.hostId] = {
-                    interested: matchResult.guestInterestLevel,
-                    lastUpdated: matchResult.lastInterestUpdate
-                };
-                return map;
-            }, {} as InterestMapping);
-    }, [data.matchResults]);
-
     const location = useLocation();
 
     React.useEffect(() => {
@@ -239,6 +92,8 @@ export const AdminGuestView = () => {
         }
     }, [location.pathname, location.search]);
 
+
+
     const guestResponsesByKey = data.guestResponses
         .filter((r: GuestResponse) => {
             return r.guestId === guest.id;
@@ -250,13 +105,6 @@ export const AdminGuestView = () => {
             )
             return prev;
         }, new Map<string, string>());
-
-    const questionsByKey = data.guestQuestions
-        .reduce<Map<string, GuestQuestion>>((prev: Map<string, GuestQuestion>, cur: GuestQuestion) => {
-            prev.set(cur.questionKey, cur);
-            return prev;
-        }, new Map<string, GuestQuestion>());
-
 
 
     const parentingResponse = guestResponsesByKey.get('parenting_guest') as string;
@@ -286,169 +134,10 @@ export const AdminGuestView = () => {
         }
     }
 
-    const FailCell = (props: { value: string }) => <div>
-        <div style={{ padding: '0px 5px', textAlign: 'center' }}>
-            <FontAwesomeIcon
-                icon={faTimesCircle}
-                style={{ 'color': 'red' }}
-            />
-
-        </div>
-        <div style={{ padding: '0px 3px', textAlign: 'center' }}>{props.value}</div>
-    </div>;
-
-    const SuccessCell = (props: { value: string }) => <div style={{ padding: '0px 3px', textAlign: 'center' }}>
-        {props.value}
-    </div>
-
-    const MatchTable = (props: { tableName: string, hostList: Array<Host>, displayInterested: boolean, allowClick: boolean }) => {
-        return (
-            <AdminGuestStyle.AdminMatchHolders>
-                <Paper className={classes.paper}>
-                    <Typography component='h2' align='left'>{props.tableName}</Typography>
-                    <Table className={classes.table} aria-label={`${props.tableName.toLowerCase()} table`}>
-
-                        <AdminGuestStyle.TableHeaderRow className={classes.tableHeader}>
-                            <AdminGuestStyle.TableHeaderLabel className={classes.matchingHeaderCell}>Name</AdminGuestStyle.TableHeaderLabel >
-                            <AdminGuestStyle.TableHeaderLabel className={classes.matchingHeaderCell}>Address</AdminGuestStyle.TableHeaderLabel >
-                            {
-                                data.hostQuestions.map((q: HostQuestion, index: number) => {
-                                    return (
-                                        <AdminGuestStyle.TableHeaderLabel className={classes.matchingHeaderCell} key={index}>{q.displayName}</AdminGuestStyle.TableHeaderLabel>
-                                    );
-                                })
-                            }
-                        </AdminGuestStyle.TableHeaderRow>
-                        <TableBody>
-                            {
-                                props.hostList.map(
-                                    (host: Host, index: number) => <><AdminGuestStyle.TableRow key={index} className={index % 2 === 0 ? classes.tableRowEven : classes.tableRowOdd}>
-                                        <AdminGuestStyle.TableCell onClick={
-                                            props.allowClick
-                                                ? () => {
-                                                    console.log(`AdminGuestView:MatchTable: guestId = ${guestId}`);
-                                                    console.log(`AdminGuestView:MatchTable: host.id = ${host.id}`);
-                                                    history.push(`/hosthome/guests/${guestId}/matches/${host.id}`)
-                                                }
-                                                : () => { }
-                                        }>
-                                            <AdminGuestStyle.HostMatchClick>
-                                                {host.name}
-                                            </AdminGuestStyle.HostMatchClick>
-                                        </AdminGuestStyle.TableCell>
-
-
-                                        <AdminGuestStyle.TableCell>{host.address}</AdminGuestStyle.TableCell>
-                                        {
-                                            data.hostQuestions.map((q: HostQuestion, index: number) => {
-                                                return (
-                                                    <AdminGuestStyle.TableCell key={index}>
-                                                        {
-                                                            (() => {
-
-                                                                if (q.questionKey === 'duration_of_stay') {
-
-                                                                    if (!hostQuestionsFailed.has(host.id)) {
-                                                                        return <SuccessCell value={hostTypeDisplay(host.type)} />;
-                                                                    }
-
-                                                                    const failedQuestionsForHost = hostQuestionsFailed.get(host.id) as Array<number>;
-                                                                    const isFailed = failedQuestionsForHost.indexOf(q.id) >= 0;
-
-                                                                    return isFailed
-                                                                        ? <FailCell value={hostTypeDisplay(host.type)} />
-                                                                        : <SuccessCell value={hostTypeDisplay(host.type)} />;
-
-                                                                }
-
-                                                                if (q.questionKey === 'hosting_amount') {
-
-                                                                    if (!hostQuestionsFailed.has(host.id)) {
-                                                                        return <SuccessCell value={host.hostingAmount.toString()} />;
-                                                                    }
-
-                                                                    const failedQuestionsForHost = hostQuestionsFailed.get(host.id) as Array<number>;
-                                                                    const isFailed = failedQuestionsForHost.indexOf(q.id) >= 0;
-
-                                                                    return isFailed
-                                                                        ? <FailCell value={host.hostingAmount.toString()} />
-                                                                        : <SuccessCell value={host.hostingAmount.toString()} />;
-
-                                                                }
-
-
-                                                                const response = data.hostResponses
-                                                                    .find((hr: HostResponse) => hr.hostId == host.id && hr.questionId == q.id);
-
-                                                                if (!response) {
-                                                                    return 'Not answered';
-                                                                }
-                                                                return response
-                                                                    .responseValues
-                                                                    .map((rvId: number, index: number) => {
-
-                                                                        const rv = data.responseValues
-                                                                            .find((rv: ResponseValue) => rv.id === rvId);
-                                                                        if (!rv) {
-                                                                            throw new Error(`Unknown response value ID: ${rvId}`);
-                                                                        }
-
-
-                                                                        if (!hostQuestionsFailed.has(host.id)) {
-                                                                            return <SuccessCell value={rv.text} />;
-                                                                        }
-
-
-                                                                        const failedQuestionsForHost = hostQuestionsFailed.get(host.id) as Array<number>;
-
-
-                                                                        const isFailed = failedQuestionsForHost.indexOf(q.id) >= 0;
-
-
-                                                                        return isFailed
-                                                                            ? <FailCell key={index} value={rv.text} />
-                                                                            : <SuccessCell key={index} value={rv.text} />;
-
-                                                                    })
-                                                            })()
-                                                        }
-                                                    </AdminGuestStyle.TableCell>
-                                                );
-                                            })
-                                        }
-                                    </AdminGuestStyle.TableRow >
-                                        {
-                                            props.displayInterested && interestByHostId[host.id].interested === GuestInterestLevel.Interested
-                                                ? <TableRow key={index}>
-                                                    <TableCell
-                                                        colSpan={data.hostQuestions.length + 2}
-                                                        style={{
-                                                            backgroundColor: '#80e27e'
-                                                        }}
-                                                    >
-                                                        {`Guest indicated interest at ${interestByHostId[host.id].lastUpdated.toLocaleString()}`}
-                                                    </TableCell>
-                                                </TableRow>
-                                                : null
-                                        }
-                                    </>
-                                )
-                            }
-                        </TableBody>
-                    </Table>
-                </Paper>
-            </AdminGuestStyle.AdminMatchHolders>
-
-        );
-    }
-
-
 
     return (
         <React.Fragment>
             <AdminGuestStyle.MainHolder>
-
-                {/* <Grid container spacing={3}> */}
 
                 {/* Page Title */}
                 <AdminGuestStyle.MainHeader>
@@ -564,6 +253,7 @@ export const AdminGuestView = () => {
                         </Paper>
                     </AdminGuestStyle.NoWrapHolder>
                 </AdminGuestStyle.AdminMatchHolders>
+
 
 
                 <MatchTable tableName='Matched' hostList={matched} allowClick={true} displayInterested={true} />
